Added framework for developing and testing stuff.
In MenuBars added an if statement which checks developing in FTCauto.
If true, a new Menu is added with testing methods.  Currently contains get distance

In FTCauto changed showCredits to developing, added color warningRed, and if statement that adds "Developing!" to the upper left corner.
Also started to add the testing for get distance method in Point.